+++ lucabook/CHANGELOG.md
@@ -1,3 +1,7 @@
+## LucaBook master
+
+* add import option, 'x-customer' in YAML settings.
+
 ## LucaBook 0.2.23
 
 * initial implement `luca-book journals list -c --customer`, grouping with x-customer header